requirements.txt
streamlit==1.32.2
google-generativeai==0.3.2
python-dotenv==1.0.1
pandas==2.2.1
PyPDF2==3.0.1
requests==2.31.0
beautifulsoup4==4.12.3
gTTS==2.5.1
moviepy==1.0.3
whisper==1.1.10
ffmpeg-python==0.2.0
TTS==0.21.1
indic-trans==2.4.1
srt==3.5.3
